Split file services up in two since more methods will be added for upcoming functionality
