+++ src/Enterspeed.Cli/Services/FileService/IDeploymentPlanFileService.cs
@@ -1,0 +1,10 @@
+﻿using Enterspeed.Cli.Services.FileService.Models;
+
+namespace Enterspeed.Cli.Services.FileService
+{
+    public interface IDeploymentPlanFileService
+    {
+        DeploymentPlanProperties GetDeploymentPlan(string filename = "deploymentplan.json");
+        void UpdateDeploymentPlan(string schemaAlias, int version);
+    }
+}
